Remove the splunk package from /tmp after install
If we're running on a Debian family OS, we have to download the file to
a temporary location before we can pass it into dpkg.  This changes the
existing behavior to rm that file after it's installed and also changes
the name of the temporary file to indicate that it was created by
puppet.