Move puppet_manage_package script to splunk basedir
The script to install splunk packages was being stored in root's home
directory, presumably because at the time it's created, the splunk
packages haven't been installed and so they haven't created /opt/splunk.
This commit puts the splunk basedir under puppet management if it's
managing the packages and then puts the puppet_manage_package script in
that directory.