Set home directory in admin password change script
When the splunk process runs as a non root user and the script runs as
root (such as running from Puppet). The splunk command line fails and
complains that it does not have permissions to create the
authentication token in roots home directory.

Since this script should be a one of, setting to root to /tmp and
deleting the token after password has changed should be enough and safe.

@yasn77